Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pymongo==4.10.1
pydantic==2.9.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
youtube-transcript-api==0.6.2
tensorflow==2.20.0
